extract appointment form into a separate component
